Remove device frame body/bezels; clip at corners and cutouts (step 4.2)
- Replace DeviceFramePainter/DeviceFrameWidget with ScreenClipPainter/ScreenClipWidget
- Painter clips canvas to rounded screen corners via profile.screenCornerRadius,
  fills black, then further restricts clip to exclude cutout region
- screenRectForSize now returns full painter bounds (no bezel deflation)
- Remove body rect, bezel constants, classic decorations (speaker/home button)
- ScreenClipWidget simplified: CustomPaint with child, no LayoutBuilder/Stack/Positioned
- Update all imports and tests; remove bezel-size assertion tests
