+++ test/frame/screen_clip_painter_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:bezel/src/devices/device_database.dart';
 import 'package:bezel/src/devices/device_profile.dart';
 import 'package:bezel/src/devices/screen_cutout.dart';
-import 'package:bezel/src/frame/device_frame_painter.dart';
+import 'package:bezel/src/frame/screen_clip_painter.dart';
 
 void main() {
   // A helper that builds a DeviceProfile for a given cutout.
@@ -25,87 +25,16 @@ void main() {
     );
   }
 
-  group('DeviceFramePainter.screenRectForSize', () {
-    test('screen rect is contained within painter bounds', () {
+  group('ScreenClipPainter.screenRectForSize', () {
+    test('screen rect equals full painter bounds', () {
       const painterSize = Size(390, 844);
       for (final profile in DeviceDatabase.all) {
-        final rect = DeviceFramePainter.screenRectForSize(
+        final rect = ScreenClipPainter.screenRectForSize(
           painterSize,
           profile,
           DeviceOrientation.portrait,
         );
-        expect(rect.left, greaterThanOrEqualTo(0));
-        expect(rect.top, greaterThanOrEqualTo(0));
-        expect(rect.right, lessThanOrEqualTo(painterSize.width));
-        expect(rect.bottom, lessThanOrEqualTo(painterSize.height));
-      }
-    });
-
-    test('NoCutout portrait has larger top and bottom bezels than sides', () {
-      const painterSize = Size(390, 844);
-      final profile = makeProfile(cutout: const NoCutout());
-      final rect = DeviceFramePainter.screenRectForSize(
-        painterSize,
-        profile,
-        DeviceOrientation.portrait,
-      );
-      final topBezel = rect.top;
-      final bottomBezel = painterSize.height - rect.bottom;
-      final sideBezel = rect.left;
-
-      expect(topBezel, greaterThan(sideBezel));
-      expect(bottomBezel, greaterThan(topBezel));
-    });
-
-    test(
-      'NoCutout landscape has larger left and right bezels than top/bottom',
-      () {
-        const painterSize = Size(844, 390);
-        final profile = makeProfile(cutout: const NoCutout());
-        final rect = DeviceFramePainter.screenRectForSize(
-          painterSize,
-          profile,
-          DeviceOrientation.landscape,
-        );
-        final leftBezel = rect.left;
-        final rightBezel = painterSize.width - rect.right;
-        final topBezel = rect.top;
-
-        expect(leftBezel, greaterThan(topBezel));
-        expect(rightBezel, greaterThan(leftBezel));
-      },
-    );
-
-    test('modern cutout profiles have uniform small bezels in portrait', () {
-      const painterSize = Size(390, 844);
-      final modernCutouts = <ScreenCutout>[
-        const DynamicIslandCutout(size: Size(37, 12), topOffset: 14),
-        const PunchHoleCutout(diameter: 11, topOffset: 13),
-        const NotchCutout(size: Size(150, 30)),
-      ];
-      for (final cutout in modernCutouts) {
-        final profile = makeProfile(cutout: cutout);
-        final rect = DeviceFramePainter.screenRectForSize(
-          painterSize,
-          profile,
-          DeviceOrientation.portrait,
-        );
-        // All four bezels should be equal for modern cutout profiles.
-        expect(
-          rect.left,
-          equals(painterSize.width - rect.right),
-          reason: '${cutout.runtimeType} left/right bezels differ',
-        );
-        expect(
-          rect.top,
-          equals(painterSize.height - rect.bottom),
-          reason: '${cutout.runtimeType} top/bottom bezels differ',
-        );
-        expect(
-          rect.left,
-          equals(rect.top),
-          reason: '${cutout.runtimeType} side/top bezels differ',
-        );
+        expect(rect, equals(Offset.zero & painterSize));
       }
     });
 
@@ -113,7 +42,7 @@ void main() {
       for (final profile in DeviceDatabase.all) {
         for (final orientation in DeviceOrientation.values) {
           const size = Size(400, 800);
-          final rect = DeviceFramePainter.screenRectForSize(
+          final rect = ScreenClipPainter.screenRectForSize(
             size,
             profile,
             orientation,
@@ -133,7 +62,7 @@ void main() {
     });
   });
 
-  group('DeviceFramePainter rendering', () {
+  group('ScreenClipPainter rendering', () {
     // Verify that every real database profile renders without throwing.
     for (final profile in DeviceDatabase.all) {
       testWidgets('renders ${profile.id} without throwing', (tester) async {
@@ -143,7 +72,7 @@ void main() {
               width: 390,
               height: 844,
               child: CustomPaint(
-                painter: DeviceFramePainter(
+                painter: ScreenClipPainter(
                   profile: profile,
                   orientation: DeviceOrientation.portrait,
                 ),
@@ -163,7 +92,7 @@ void main() {
             width: 390,
             height: 844,
             child: CustomPaint(
-              painter: DeviceFramePainter(
+              painter: ScreenClipPainter(
                 profile: profile,
                 orientation: DeviceOrientation.portrait,
               ),
@@ -184,7 +113,7 @@ void main() {
             width: 844,
             height: 390,
             child: CustomPaint(
-              painter: DeviceFramePainter(
+              painter: ScreenClipPainter(
                 profile: profile,
                 orientation: DeviceOrientation.landscape,
               ),
